requirements.txt
Django~=3.1.5
gunicorn==19.7.1
whitenoise==3.3.1
django-cors-headers==3.7.0
djangorestframework==3.12.0
requests==2.25.1
six==1.15.0
django-utils-six==2.0